requirements.txt
django-oscar>=2.0,<2.1
coverage==4.5.4
django-webtest==1.9.7
pytest-django>=3.7,<3.8
tox>=2.9,<3.15
sorl-thumbnail>=12.4.1,<12.5
